requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.4
xgboost>=2.0
imbalanced-learn>=0.12
matplotlib>=3.8
plotly>=5.20
pyyaml>=6.0
joblib>=1.3
streamlit>=1.34
